requirements.txt
# Google Cloud SDK and BigQuery
google-cloud-bigquery>=3.13.0
google-cloud-storage>=2.10.0
google-auth>=2.23.0

# Data processing
pandas>=2.1.0
pyarrow>=14.0.0

# Configuration
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.9.0
flake8>=6.1.0
mypy>=1.5.0

# Optional: Airflow (uncomment if needed)
# apache-airflow>=2.7.0
# apache-airflow-providers-google>=10.8.0
